mysite/requirements.txt
Django>=3.2,<3.3
djangorestframework>=3.9.2
djangorestframework_simplejwt>=5.1.0
django-cors-headers>=3.11.0
django-environ>=0.4.5
gunicorn>=20.1.0
psycopg2>=2.9.3
pillow>=9.0.0
requests>=2.27.0